+++ web/apps/photos/src/components/Collections/AllCollections/index.tsx
@@ -1,8 +1,8 @@
 import { CollectionsSortOptions } from "@/new/photos/components/CollectionsSortOptions";
 import { FilledIconButton } from "@/new/photos/components/mui";
 import { SlideUpTransition } from "@/new/photos/components/mui/SlideUpTransition";
-import type { CollectionSummary } from "@/new/photos/types/collection";
-import { CollectionsSortBy } from "@/new/photos/types/collection";
+import type { CollectionSummary } from "@/new/photos/services/collection/ui";
+import { CollectionsSortBy } from "@/new/photos/services/collection/ui";
 import { FlexWrapper, FluidContainer } from "@ente/shared/components/Container";
 import Close from "@mui/icons-material/Close";
 import {
