+++ web/apps/photos/src/utils/collection.ts
@@ -8,11 +8,11 @@ import {
     CollectionType,
 } from "@/media/collection";
 import { ItemVisibility } from "@/media/file-metadata";
-import { getAllLocalFiles, getLocalFiles } from "@/new/photos/services/files";
 import type {
     CollectionSummaries,
     CollectionSummaryType,
-} from "@/new/photos/types/collection";
+} from "@/new/photos/services/collection/ui";
+import { getAllLocalFiles, getLocalFiles } from "@/new/photos/services/files";
 import { EnteFile } from "@/new/photos/types/file";
 import { SUB_TYPE } from "@/new/photos/types/magicMetadata";
 import { safeDirectoryName } from "@/new/photos/utils/native-fs";
